refactor: extract monorepo analysis into analyzer/monorepo.rs (#190)
Moves CrossModuleViolation, MonorepoResult, and audit_modules out of
analyzer/mod.rs into a dedicated submodule. Five tests move with them.
